pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stocklab"
version = "0.1.0"
description = "AI-powered stock intelligence dashboard for Indonesia market"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.44.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "APScheduler>=3.10.4",
  "yfinance>=0.2.54"
]

[tool.setuptools.packages.find]
include = ["stocklab*"]
